test(codecs): port TestTrie + fix 2 production bugs in trie reader/builder
Sprint 57, GOC-4113. Fixes: lucene103_trie_reader.go uses
int64(uint64(encodedFP) >> 2) for logical shift; lucene103_trie_builder.go
adds bitIndex&63 mask to shift count. 4/4 tests pass.

+++ codecs/lucene103_trie_reader.go
@@ -274,7 +274,7 @@ func (r *TrieReader) loadSingleChildNode(n *TrieNode, fp int64, sign, term int, 
 		return err
 	}
 	encodedFP := encodedRaw & trieBytesMinus1Mask[encodedOutputFPBytesMinus1]
-	n.OutputFP = encodedFP >> 2
+	n.OutputFP = int64(uint64(encodedFP) >> 2)
 	n.HasTerms = encodedFP&trieNonLeafNodeHasTerms != 0
 	if encodedFP&trieNonLeafNodeHasFloor != 0 {
 		n.FloorDataFP = offset + int64(encodedOutputFPBytesMinus1) + 1
@@ -304,7 +304,7 @@ func (r *TrieReader) loadMultiChildrenNode(n *TrieNode, fp int64, term int, term
 			l = v
 		}
 		encodedFP := l & trieBytesMinus1Mask[encodedOutputFPBytesMinus1]
-		n.OutputFP = encodedFP >> 2
+		n.OutputFP = int64(uint64(encodedFP) >> 2)
 		n.HasTerms = encodedFP&trieNonLeafNodeHasTerms != 0
 		if encodedFP&trieNonLeafNodeHasFloor != 0 {
 			offset := fp + 4 + int64(encodedOutputFPBytesMinus1)
